+++ RedPandaIDE/translations/RedPandaIDE_zh_TW.ts
@@ -1241,7 +1241,7 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="+2894"/>
+        <location line="+2906"/>
         <source>Print Document</source>
         <translation type="unfinished"></translation>
     </message>
@@ -1279,7 +1279,7 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="-5397"/>
+        <location line="-5409"/>
         <location line="+505"/>
         <source>Error Load File</source>
         <translation type="unfinished"></translation>
@@ -2422,7 +2422,7 @@
     </message>
     <message>
         <location line="+21"/>
-        <source>Search Actions</source>
+        <source>Filter Actions</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
@@ -4019,7 +4019,7 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="+497"/>
+        <location line="+499"/>
         <location line="+2252"/>
         <source>Files</source>
         <translation type="unfinished"></translation>
@@ -4157,8 +4157,8 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="-1037"/>
-        <location line="+1076"/>
+        <location line="-1039"/>
+        <location line="+1078"/>
         <location line="+1505"/>
         <location filename="../mainwindow.cpp" line="+5886"/>
         <source>Search</source>
@@ -4281,7 +4281,7 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="-1788"/>
+        <location line="-1790"/>
         <source>File</source>
         <translation type="unfinished"></translation>
     </message>
@@ -4303,7 +4303,7 @@
     </message>
     <message>
         <location line="+55"/>
-        <location line="+161"/>
+        <location line="+163"/>
         <location filename="../mainwindow.cpp" line="+4324"/>
         <location line="+1"/>
         <location line="+1"/>
@@ -4824,7 +4824,17 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="-449"/>
+        <location line="+267"/>
+        <source>Code Completion</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="+3"/>
+        <source>Ctrl+Shift+/</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <location line="-719"/>
         <source>Find references</source>
         <translation type="unfinished"></translation>
     </message>
@@ -6654,7 +6664,7 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="-3072"/>
+        <location filename="../mainwindow.ui" line="-3074"/>
         <source>Move Caret</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8218,12 +8228,12 @@
     </message>
     <message>
         <location filename="../compiler/compilerinfo.cpp" line="+92"/>
-        <location line="+347"/>
+        <location line="+349"/>
         <source>Code Generation</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="-261"/>
+        <location line="-263"/>
         <source>Enable use of specific instructions (-mx)</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8283,7 +8293,7 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="+5"/>
+        <location line="+7"/>
         <source>Do not use standard system libraries (-nostdlib)</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8298,12 +8308,12 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="-40"/>
+        <location line="-42"/>
         <source>Generate debugging information (-g3)</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="+36"/>
+        <location line="+38"/>
         <source>Use pipes instead of temporary files during compilation (-pipe)</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8421,7 +8431,7 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location filename="../compiler/compilerinfo.cpp" line="-105"/>
+        <location filename="../compiler/compilerinfo.cpp" line="-107"/>
         <source>C++ Language standard (-std)</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8491,7 +8501,7 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="+248"/>
+        <location line="+250"/>
         <source>Processor (-m)</source>
         <translation type="unfinished"></translation>
     </message>
@@ -8511,12 +8521,12 @@
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="-262"/>
+        <location line="-263"/>
         <source>Stack Size</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <location line="+263"/>
+        <location line="+264"/>
         <source>Use movc instead of movx to read from external ram</source>
         <translation type="unfinished"></translation>
     </message>