+++ NEWS.md
@@ -165,6 +165,7 @@ Red Panda C++ Version 2.27
   - change: Set "Ctrl+G" as the shortcut for "Goto page..."
   - change: Set "Ctrl+B" as the shortcut for "Toggle Bookmark"
   - fix: Fail to evaluate expressions if macro can't be expanded.
+  - enhancement: New menu item "Code completion".
     
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.